requirements.txt
SpeechRecognition>=3.8.1
pyttsx3>=2.90
pyautogui>=0.9.54
pyinstaller>=4.7  # Optional - only if creating executables
pillow>=11.1.0  
numpy>=2.2.1
opencv-contrib-python>=4.10.0.84
screeninfo>=0.8.1
moviepy>=2.1.2